tasks4/pyproject.toml
[project]
name = "tasks4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Brandon Tamir", email = "brandon.tamir@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai>=2.7.2",
    "python-dotenv>=1.2.1",
]

[project.scripts]
tasks4 = "tasks4:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
